6th commit(Fixed latest issues) (Notification creation moved into Application Class)
