+++ app/src/main/java/com/example/movieapp/data/repository/MovieNetworkRepositoryImpl.kt
@@ -13,6 +13,7 @@ import com.example.movieapp.domain.utils.Resource
 import kotlinx.coroutines.async
 import kotlinx.coroutines.awaitAll
 import kotlinx.coroutines.coroutineScope
+import kotlinx.coroutines.runBlocking
 
 class MovieNetworkRepositoryImpl(
     private val networkClient: NetworkClient,
@@ -86,6 +87,10 @@ class MovieNetworkRepositoryImpl(
                 async { requestActorsDetails(actor) }
             }
         }.awaitAll()
+
+        runBlocking {
+
+        }
 
         val errors = result.filterIsInstance<Resource.Error<ActorsResponseDto>>()
