+++ app/src/main/java/com/example/movieapp/presentation/ui/movieDetails/MovieDetailsFragment.kt
@@ -1,10 +1,12 @@
 package com.example.movieapp.presentation.ui.movieDetails
 
+import android.os.Build
 import androidx.core.view.isVisible
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.fragment.navArgs
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.movieapp.R
+import com.example.movieapp.app.App.Companion.WARNING
 import com.example.movieapp.databinding.MovieDetailsFragmentBinding
 import com.example.movieapp.domain.model.MovieDetailsModel
 import com.example.movieapp.domain.utils.isOnline
@@ -61,13 +63,19 @@ class MovieDetailsFragment :
 
     private fun initListeners() {
         binding.watchLater.setOnClickListener {
-            val action =
-                MovieDetailsFragmentDirections
-                    .actionMovieDetailsFragmentToItemListDialogFragment(
-                        binding.titleTV.text.toString(),
-                        safeArgs.id
-                    )
-            findNavController().navigate(action)
+            val title = binding.titleTV.text
+            if (!title.isNullOrEmpty() && (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O)) {
+                val action =
+                    MovieDetailsFragmentDirections
+                        .actionMovieDetailsFragmentToItemListDialogFragment(
+                            title.toString(),
+                            safeArgs.id
+                        )
+                findNavController().navigate(action)
+            }
+            else {
+                makeToastMessage(WARNING)
+            }
         }
 
         binding.backButton.setOnClickListener {
