+++ app/src/main/res/layout/fragment_item_list_dialog_list_dialog.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:clipToPadding="false"
-    tools:context=".presentation.ui.datePickerDialog.ScheduleNotificationDialogFragment">
+    tools:context=".presentation.ui.scheduleNotification.ScheduleNotificationDialogFragment">
 
     <TimePicker
         android:id="@+id/timePicker"
